Prepare game for build (fixes some file location issues)
